car-price-project/requirements.txt
streamlit
pandas
numpy
scikit-learn
xgboost
torch
matplotlib
Pillow
